+++ src/Pipelinez.AzureServiceBus/Client/IAzureServiceBusSender.cs
@@ -1,0 +1,8 @@
+using Azure.Messaging.ServiceBus;
+
+namespace Pipelinez.AzureServiceBus.Client;
+
+internal interface IAzureServiceBusSender : IAsyncDisposable
+{
+    Task SendMessageAsync(ServiceBusMessage message, CancellationToken cancellationToken);
+}