+++ src/Pipelinez.AzureServiceBus/Source/AzureServiceBusSettlementAction.cs
@@ -1,0 +1,9 @@
+namespace Pipelinez.AzureServiceBus.Source;
+
+internal enum AzureServiceBusSettlementAction
+{
+    Complete,
+    Abandon,
+    DeadLetter,
+    None
+}